+++ CHANGELOG.md
@@ -7,4 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Build system
+
 [Unreleased]: https://github.com/sonro/spat/compare/23a8e0d382a89f6dff8a205ee0fd0c2d449df6e7...HEAD